% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{anonymize}
\alias{anonymize}
\title{Anonymize}
\usage{
anonymize(x, algo = "crc32")
}
\arguments{
\item{x}{String to anonymize.}

\item{algo}{Algorithm to anonymize id. Default: crc32.}
}
\description{
Create anonymized id. Note: function is still on alpha stage.
TODO: create id variable from source and date_time
TODO: salt aka add some noise characters to variable
}
\details{
# Original function and its use
anonymize <- function(x, algo="crc32"){
 unq_hashes <- vapply(unique(x), function(object) digest(object, algo=algo), FUN.VALUE="", USE.NAMES=TRUE)
 unname(unq_hashes[x])
}
cols_to_mask <- c("name", "address", "postal_code")
SURV_ORG[, cols_to_mask := lapply(.SD, anonymize), .SDcols = cols_to_mask, with = FALSE][]
}
\examples{
anonymize(0:10, algo = "crc32")
}
\keyword{anonymize}
\keyword{data}
