% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_essentials.R
\name{api_essentials}
\alias{api_essentials}
\title{api_essentials}
\usage{
api_essentials(soap_action, username, password, quest_id, security_lock, x,
  send_duplicate = TRUE)
}
\arguments{
\item{username}{Username for Questback Essentials.}

\item{password}{Password for Questback Essentials integration.}

\item{x}{the file name to save the dataframe.}

\item{action}{API action: TestConnection / AddRespondentsData / AddEmailInvitees.}

\item{sep}{Choose separator. Default: ";"}
}
\description{
API to POST/GET Questback Essentials survey information
}
\examples{
api_essentials("TestConnection", "Your username", "Your password")
}
\keyword{API}
\keyword{survey}
