% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_plot.R
\name{basic_plot}
\alias{basic_plot}
\title{Custom basic plot}
\usage{
basic_plot(data, select)
}
\arguments{
\item{data}{dataframe}

\item{select}{select columns}

\item{col}{select columns?}
}
\description{
Make a custom basic plot.
}
\examples{
basic_plot(data, select = c(1:10, 13))
}
\keyword{plot}
