% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound.R
\name{bound}
\alias{bound}
\title{bound}
\usage{
bound(x, lower, upper)
}
\arguments{
\item{x}{Choose variable from which to calculate.}

\item{lower}{Lower bound to cut the values to.}

\item{upper}{Upper bound to limit the values to.}
}
\description{
Bound the variable by lower and upper limits
}
\examples{
bound(-10, 0, 10)
bound(c(-10, 0, 3, 10, 100, "r"), 0, 10)
sapply(c(-10, 0, 100, NULL, NA), bound, 0, 10)
}
\keyword{bound,}
\keyword{cut}
\keyword{limit,}
