% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand_cols.R
\name{brand_cols}
\alias{brand_cols}
\title{Function to extract Brand colors as hex codes}
\usage{
brand_cols(...)
}
\arguments{
\item{...}{Character names of brand_colors}
}
\description{
Function to extract Brand colors as hex codes
}
