% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_area.R
\name{calc_area}
\alias{calc_area}
\title{calc_area}
\usage{
calc_area(x, select = "Alue")
}
\arguments{
\item{x}{Choose variable from which to calculate.}

\item{output}{Choose variable(s) to return e.g. Maakunta/Maakunnan.nimi/Kuntaryhma/Kuntaryhman.nimi/Seutukunta/Seutukunnan.nimi/Suuralue/Suuralueen.nimi/Alue)}
}
\description{
Calculate area classification e.g. (Etelä-Suomi - Pohjois-Suomi) from first three postnumber characters (000-999)
}
\examples{
calc_area("001")
}
\keyword{area,}
\keyword{calculation,}
\keyword{classification}
