% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_birth.R
\name{calc_birth}
\alias{calc_birth}
\title{calc_birth}
\usage{
calc_birth(string)
}
\arguments{
\item{string}{Choose string from which to calculate.}
}
\description{
Calculate birth date from social security number
}
\examples{
calc_birth("123456-123S")
}
\keyword{birth,}
\keyword{date}
\keyword{security,}
\keyword{social}
