% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dummy.R
\name{calc_dummy}
\alias{calc_dummy}
\title{calc_dummy}
\usage{
calc_dummy(vars)
}
\arguments{
\item{vars}{Choose variables from which to calculate.}
}
\description{
Calculate 0-1 dummy variable from 2-n variables.
}
\examples{
calc_dummy("", "1", "")
}
\keyword{calculate}
\keyword{dummy}
