% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gender.R
\name{calc_gender}
\alias{calc_gender}
\title{calc_gender}
\usage{
calc_gender(string)
}
\arguments{
\item{string}{Choose string from which to calculate.}
}
\description{
Calculate gender from social security number
}
\examples{
calc_gender("123456-123S")
}
\keyword{gender}
\keyword{security,}
\keyword{social}
