% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nps.R
\name{calc_nps}
\alias{calc_nps}
\title{calc_nps}
\usage{
calc_nps(var = NA)
}
\arguments{
\item{var}{Choose variable from which to calculate.}
}
\description{
Calculate NPS from survey
}
\examples{
calc_nps(0)
}
\keyword{birth,}
\keyword{date}
\keyword{security,}
\keyword{social}
