% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tol.R
\name{calc_tol}
\alias{calc_tol}
\title{calc_tol}
\usage{
calc_tol(x)
}
\arguments{
\item{x}{Choose variable from which to calculate.}
}
\description{
Calculate industry upper level classification (A-X) from lower levels (00-99)
}
\examples{
calc_tol("00")
}
\keyword{calculation,}
\keyword{classification}
\keyword{industry,}
