% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cantype.R
\name{cantype}
\alias{cantype}
\title{Undescribe}
\usage{
cantype(x, ...)
}
\arguments{
\item{Undescribe}{}
}
\value{
Undescribe
}
\description{
Undescribe
}
\examples{
(dat_logi <- as.character(sample(c(TRUE, FALSE), 10, TRUE)))
cantype(dat_logi)
(dat_int <- as.character(sample(5, 10, TRUE)))
cantype(dat_int)
(dat_num <- as.character(rnorm(10)))
cantype(dat_num)
}

