% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.factor.levels.R
\name{change.factor.levels}
\alias{change.factor.levels}
\title{Change the levels of a factor}
\usage{
change.factor.levels(input.factor, old.list, new.list)
}
\description{
Change the levels of a factor
}
