\name{\%+\%}
\alias{\%+\%}
\title{Concatenate two strings}
\usage{
s1 \%+\% s2
}
\arguments{
  \item{s1}{First String}

  \item{s2}{Second String}
}
\value{
\code{paste(s1,s2,sep="")}
}
\description{
Paste two strings together without separation.
}
\examples{
cat("Hello" \%+\% "World")
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

