% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.constant.metrics.R
\name{check.constant.metrics}
\alias{check.constant.metrics}
\title{Remove metrics that are constant}
\usage{
check.constant.metrics(dataset, indep)
}
\arguments{
\item{dataset}{a data frame for data}

\item{indep}{a characters or a vector of characters for independent variables}
}
\description{
This function makes life simple by checking and removing constant metrics.
}
\keyword{Constant}
