% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_company_names.R
\name{clean_company_names}
\alias{clean_company_names}
\title{clean_company_names}
\usage{
clean_company_names(x)
}
\arguments{
\item{x}{Choose string which to clean.}
}
\description{
Clean string from Finnish company names
}
\examples{
clean_company_names("lklkdf Nokia Oyj lkjadf ef Kämmekkäsäätiö")
}
\keyword{clean,}
\keyword{company,}
\keyword{name}
\keyword{string,}
