% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_emails.R
\name{clean_emails}
\alias{clean_emails}
\title{clean_emails}
\usage{
clean_emails(x)
}
\arguments{
\item{x}{Choose string which to clean.}
}
\description{
Clean string from emails
}
\examples{
clean_emails("lklkdf sari.pukkila@pukkila.fi lkjadf ef")
}
\keyword{clean,}
\keyword{email}
\keyword{string,}
