% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ids.R
\name{clean_ids}
\alias{clean_ids}
\title{clean_ids}
\usage{
clean_ids(x)
}
\arguments{
\item{x}{Choose string which to clean.}
}
\description{
Clean string from possible id numbers
}
\examples{
clean_ids("003904A")
}
\keyword{clean,}
\keyword{strees}
\keyword{string,}
