% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_nochars.R
\name{clean_nochars}
\alias{clean_nochars}
\title{clean_nochars}
\usage{
clean_nochars(x)
}
\arguments{
\item{x}{Choose string which to clean.}
}
\description{
Clean string from nochars: carriage return, new line and tab
}
\examples{
clean_nochars("Jouni\nVatanen")
}
\keyword{clean,}
\keyword{line}
\keyword{new}
\keyword{string,}
