% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_people_names.R
\name{clean_people_names}
\alias{clean_people_names}
\title{clean_people_names}
\usage{
clean_people_names(x)
}
\arguments{
\item{x}{Choose string which to clean.}
}
\description{
Clean string from Finnish people
}
\examples{
clean_people_names("Ari, Jouni, EsaJokutie 4, pikitie 7, 00500 Helsinki")
}
\keyword{clean,}
\keyword{strees}
\keyword{string,}
