% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_post_numbers.R
\name{clean_post_numbers}
\alias{clean_post_numbers}
\title{clean_post_numbers}
\usage{
clean_post_numbers(x)
}
\arguments{
\item{x}{Choose string which to clean.}
}
\description{
Clean string from Finnish post numbers
}
\examples{
clean_street_names("Jokutie 4, pikitie 7, 00500 Helsinki")
}
\keyword{clean,}
\keyword{strees}
\keyword{string,}
