% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_street_names.R
\name{clean_street_names}
\alias{clean_street_names}
\title{clean_street_names}
\usage{
clean_street_names(x)
}
\arguments{
\item{x}{Choose string which to clean.}
}
\description{
Clean string from Finnish street names
}
\examples{
clean_street_names("Jokutie 4, pikitie 7, 00500 Helsinki")
}
\keyword{clean,}
\keyword{strees}
\keyword{string,}
