% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cns.R
\name{cns}
\alias{cns}
\title{Consecutive number sequences}
\usage{
cns(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The number of consecutive the value
}
\description{
Consecutive number sequences
}
\examples{
cns(c(TRUE, TRUE, FALSE, FALSE, TRUE, FALSE, TRUE, TRUE, TRUE))
}

