% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colC.R
\name{colC}
\alias{colC}
\title{Return the Short Form to the Long Form of the type of an objects}
\usage{
colC(x)
}
\arguments{
\item{x}{A character that is the acronym of the type of an objects}
}
\value{
A character vector of the Long Forms
}
\description{
Return the Short Form to the Long Form of the type of an objects
}
\examples{
colC("dicfidl")
}

