% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand_cols.R
\name{colorRampD}
\alias{colorRampD}
\title{Function like colorRamp, but handles better situations, where color vector is
larger than is required.}
\usage{
colorRampD(
  colors,
  n,
  bias = 1,
  space = c("rgb", "Lab"),
  interpolate = c("linear", "spline"),
  alpha = FALSE
)
}
\arguments{
\item{colors}{TBD}

\item{n}{TBD}

\item{bias}{TBD}

\item{space}{TBD}

\item{interpolate}{TBD}

\item{alpha}{TBD}
}
\description{
Function like colorRamp, but handles better situations, where color vector is
larger than is required.
}
