% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand_cols.R
\name{colorRampPaletteD}
\alias{colorRampPaletteD}
\title{Function like colorRampPalette to ramp palette unless number of colors vector
is larger than is required.}
\usage{
colorRampPaletteD(colors, ...)
}
\arguments{
\item{colors}{TBD}

\item{...}{TBD}
}
\description{
Function like colorRampPalette to ramp palette unless number of colors vector
is larger than is required.
}
