% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_plot.R
\name{corr_plot}
\alias{corr_plot}
\title{Custom correlation plot}
\usage{
corr_plot(data, select = 1:ncol(data))
}
\arguments{
\item{data}{Choose the dataframe.}

\item{select}{Choose the columns you want. Default: all.}

\item{polychoric}{Not yet implemented. TRUE = use polychoric correlations.}
}
\description{
Make a custom correlation  plot.
}
\examples{
hclust_var(data, select = c(1:10, 13), polychoric = TRUE)
}
\keyword{plot}
