% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corstars.R
\name{corstars}
\alias{corstars}
\title{corstars}
\usage{
corstars(x, type = c("pearson", "spearman"), ...)
}
\arguments{
\item{x}{data frame for which pairwise correlations are calculated.}

\item{...}{other parameters passed through.}

\item{method}{either "pearson" (the default) or "spearman".}
}
\value{
prints table with correlation coefficients and significance stars.
}
\description{
Calculate pairwise correlation coefficients and print with significance stars.
}
\details{
significance code: *: p<0.05, **: p<0.01, ***: p<0.001.
}
