% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cortab.R
\name{cortab}
\alias{cortab}
\title{cortab}
\usage{
cortab(x, y, method = c("pearson", "kendall", "spearman"), ...)
}
\arguments{
\item{x}{a data frame.}

\item{y}{the response variable.}

\item{method}{one of "pearson", "kendall" or "spearman".}

\item{...}{other parameters passed through.}
}
\value{
prints table with correlation coefficients, confidence intervals and p-values.
}
\description{
Calculate correlation coefficients of many x with one y using the function \code{cor.test}.
}
\details{
\texttt{cor.test} is used for calculation and a table with estimates, confidence intervals and p-values is returned.
}
\author{
Jutta Gamper
}
