% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxregmulti.R
\name{coxregmulti}
\alias{coxregmulti}
\title{coxregmulti}
\usage{
coxregmulti(mod)
}
\arguments{
\item{mod}{a multiple cox proportional hazards model fit by \code{coxph}.}
}
\value{
prints a table with hazard ratios, confidence intervals and p-values.
}
\description{
Return a formatted output of a multiple cox regression model.
}
\author{
Jutta Gamper
}
