% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxreguni.R
\name{coxreguni}
\alias{coxreguni}
\title{coxreguni}
\usage{
coxreguni(event, time, x, etype = NULL)
}
\arguments{
\item{event}{binary indicator of event, typically 0 or 1.}

\item{time}{follow up time}

\item{x}{a data frame containing the independent variable(s) for cox the regression.}

\item{etype}{type of event for cause specific hazard models (only for competing risks).}
}
\value{
prints tables with hazard ratios and confidence intervals.
}
\description{
Calculate simple cox regression models and return a formatted output.
}
\author{
Jutta Gamper
}
\keyword{htest}
