% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_notebook}
\alias{create_notebook}
\title{Create R project}
\usage{
create_notebook(dir = ".")
}
\arguments{
\item{dir}{character: Directory where the R project is created; current
working directory by default}
}
\description{
Creates an R project with useful configuration. If the project
contains a package, the respective options are set. The project is named
after the folder which contains it.
Used in \code{\link{createProjectSkeleton}}.
}
\examples{
\dontrun{
createProject(pkg = FALSE, dir = "./")
dir.create("tmp")
createProject(pkg = TRUE, pkgOnToplevel = FALSE, dir = "tmp")}

}
