% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create project structure}
\usage{
create_project(dir = ".", notebook = TRUE, project = TRUE)
}
\arguments{
\item{dir}{character: Directory where the file structure is created, absolute
or relative to the current working directory. The current working directory
by default.}

\item{notebook}{logical: Create notebook file.}

\item{project}{logical: Create Rproject file.}
}
\description{
: This function helps you create a project structure. This includes:
\itemize{
  \item R (folder)
  \item data (folder)
  \item doc (folder)
  \item output (folder)
  \item posted (folder)
  \item work (folder)
  \item .Rprofile
  \item .Rproj (optional; created with \code{\link{createProject}})
}
}
\examples{
create_project()
}
\keyword{project}
\keyword{skeleton}
