% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_query.R
\name{db_get_query}
\alias{db_get_query}
\title{db_get_query}
\usage{
db_get_query(sql, dsn = NULL, encoding = "utf-8", locale = locale())
}
\arguments{
\item{sql}{sql query or sql file}

\item{dsn}{if NULL then use password etc.}

\item{encoding}{database encoding}

\item{locale}{locale settings for sql file}

\item{x}{Choose string from which to calculate.}
}
\description{
Get query results and close connection safely
}
\keyword{DBI,}
\keyword{database}
\keyword{odbc,}
