% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write_table2.R
\name{db_write_table2}
\alias{db_write_table2}
\title{db_write_table2}
\usage{
db_write_table2(
  data,
  con,
  schema,
  table,
  fields = NULL,
  to_nvarchar = TRUE,
  overwrite = TRUE,
  append = !overwrite,
  unique_cols = everything(),
  to_utf16 = TRUE,
  temporary = FALSE,
  logging = TRUE,
  batch_rows = 1000,
  bulk = FALSE
)
}
\arguments{
\item{data}{Data.}

\item{con}{Connection string.}

\item{schema}{Schema name.}

\item{table}{Table name.}

\item{fields}{Named vector of column names and datatypes. If null, then
automatic fetch by DBI::dbDataType.}

\item{to_nvarchar}{Convert varchar datatype to nvarchar for easier encoding.}

\item{overwrite}{Overwrite table.}

\item{append}{Append table.}

\item{unique_cols}{Columns, which makes appended data unique. You can use
tidyselect styled helpers like starts_with(). Default: everything().}

\item{to_utf16}{Convert to UTF-16LE, which is required for proper Microsoft
SQL Server encoding.}

\item{batch_rows}{Number of rows for a batch. Default: 1000. Suggestion: 10000.}

\item{bulk}{Not yet implemented. Allows bulk insert using bcp.exe tool.}
}
\description{
Write table to database
}
\keyword{database}
\keyword{sql,}
\keyword{write,}
