% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stat.R
\name{desc_stat}
\alias{desc_stat}
\title{Create descriptive statistics table}
\usage{
desc_stat(data, .select, .labels = c(total = "N"),
  group.by.cols = NULL)
}
\arguments{
\item{data}{Choose dataframe.}

\item{.select}{Choose which columns you want to calculate distribution.}

\item{.labels}{Choose labels for table. Default: total = "N"}

\item{group.by.cols}{Group by chosen columns. Defaul: NULL}
}
\description{
Create descriptive statistics table
}
\examples{
desc_stat(data, .select = c(1:10))
}
\keyword{birth,}
\keyword{date}
\keyword{security,}
\keyword{social}
