% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Description.R
\name{Description}
\alias{Description}
\title{Create the header templete of the file}
\usage{
Description(desc = "This is a test message.", symbol = "#", sharp = 50)
}
\arguments{
\item{desc}{A character that is the description}

\item{symbol}{The symbol that is consists of the header}

\item{sharp}{The width of the header}
}
\value{
The header templete of the file
}
\description{
Create the header templete of the file
}
\examples{
Description()
}

