% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmisc.R
\name{fct_min_count}
\alias{fct_min_count}
\title{A function to unite factor levels so that no level has less than 5 occurrences}
\usage{
fct_min_count(object, min = 5, other_level = "Other", max_levels = NA)
}
\arguments{
\item{object}{A factor}

\item{min}{Minimum number of occurrences for a level}

\item{other_level}{Name of the combination level}

\item{max_levels}{Maximum number of levels in the result, including the lumped level}
}
\value{
A factor
}
\description{
Combines levels with small count so that no level has less than 5 occurrences.
}
\examples{
v <- rep(letters[1:4], c(6,5,2,1)) |> as.factor()
v
fct_min_count(v)
}
\seealso{
\code{\link[forcats:fct_lump_min]{forcats::fct_lump_min()}} lumps levels that appear fewer than min times.
}
