% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fi_industries_2008}
\alias{fi_industries_2008}
\title{Finnish industry classification TOL2008}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 89 rows and 5 columns.}
\usage{
fi_industries_2008
}
\description{
Data from Tilastokeskus about Finnish industry classification.
Data is from 2008. Data can be downloaded from:
https://www.stat.fi/meta/luokitukset/toimiala/versio.html
}
\keyword{datasets}
