% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fi_people_names}
\alias{fi_people_names}
\title{Finnish names}
\format{
A data frame with 34935 rows and 4 variables:
\describe{
  \item{name}{name of a person}
  \item{n}{popularity of the name}
  \item{gender}{Male = M, Female = F, NA in lastnames}
  \item{type}{Firstname = F, lastname = L}
}
}
\usage{
fi_people_names
}
\description{
Data from Väestörekisterikeskus about Finnish first and last names by gender
and their how popular they are. Data is from 2016. Data can be downloaded from:
https://www.avoindata.fi/data/fi/dataset/none
}
\keyword{datasets}
