% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fi_postnumbers_2016}
\alias{fi_postnumbers_2016}
\title{Finnish postnumbers 2016}
\format{
A data frame with 939 rows and 10 variables:
\describe{
  \item{postinumero}{Postnumber depricated to first three characters}
  \item{maakunta}{Area codes divided by province}
  \item{maakunnan_nimi}{Area names divided by province}
  \item{kuntaryhma}{Area codes divided by area type}
  \item{kuntaryhman_nimi}{Area names divided by area type}
  \item{seutukunta}{Area codes divided by great area}
  \item{seutukunnan_nimi}{Area names divided by great area}
  \item{suuralue}{Area codes divided by NUTS2}
  \item{suuralueen_nimi}{Area names divided by NUTS2}
  \item{alue}{Area names divided by compass point}
}
}
\usage{
fi_postnumbers_2016
}
\description{
Data from Tilastokeskus about Finnish postnumbers and their metadata.
Data is from 2016. Data can be downloaded from:
}
\keyword{datasets}
