% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_date.R
\name{format_date}
\alias{format_date}
\title{format_date}
\usage{
format_date(x, format = "\%d.\%m.\%Y")
}
\arguments{
\item{x}{a date object, typically a vector of dates.}
}
\description{
Formats dates to the finnish standard.
}
\examples{
format_date(as.Date("2020-01-01"))
}
\keyword{date}
\keyword{format,}
