% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_phone_num.R
\name{format_phone_num}
\alias{format_phone_num}
\title{format_phone_num}
\usage{
format_phone_num(x, prefix = "+358")
}
\arguments{
\item{x}{Choose string which to format.}

\item{prefix}{Choose prefix in front of the phone number. Default: "+358"}
}
\description{
Format phone number to a desired format..
}
\examples{
format_phone_num("Matti 040 000 1234")
}
\keyword{clean,}
\keyword{number}
\keyword{phone}
\keyword{string,}
