% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.automated.varclus.R
\name{get.automated.varclus}
\alias{get.automated.varclus}
\title{Automated Spearman rank correlation test}
\usage{
get.automated.varclus(dataset, indep, dep, threshold, verbose = F)
}
\arguments{
\item{dataset}{a data frame for data}

\item{indep}{a characters or a vector of characters for independent variables}

\item{dep}{a character for dependent variable}

\item{verbose}{TRUE for printing}

\item{spearman.threshold}{a numeric for a threshold of Spearman rank correlation test (default = 0.7)}
}
\description{
This function automatically select non-correlated metrics based on a Spearman rank correlation test. To do so, we start from the pair of the strongest correlated metrics. Since these two metrics can be linearly predicted with each other, one of these two metrics must be removed while selecting the other. We select the metric that has the lowest Spearman correlation coefficient with the other metrics that are not in the pair. We repeat this process until all metrics have their Spearman correlation coefficient below a threshold value (default = 0.7).
}
\keyword{Spearman}
