% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.color.boxplot.R
\name{get.color.boxplot}
\alias{get.color.boxplot}
\title{Get Colorful Boxplots Function}
\usage{
get.color.boxplot(
  input.data,
  melted = F,
  xlab = "",
  ylab = "",
  fill.color = F,
  range.y = NULL,
  range.x = NULL,
  reverse.x = T,
  reverse.y = T,
  sort.x = F,
  detailed.y.ticks = F,
  y.tick.div = 1,
  text.size = 20,
  facet = F,
  width.size = 10,
  height.size = 7.5,
  file.name = "plot-output",
  color.pallete = c("#f1a340", "#f7f7f7", "#998ec3")
)
}
\arguments{
\item{input.data}{a data frame for data}

\item{melted}{}

\item{xlab}{}
}
\description{
This function makes life simple by automatically generating colorful boxplots.
}
\keyword{boxplot}
