% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.logistic.regression.R
\name{get.logistic.regression}
\alias{get.logistic.regression}
\title{Get Logistic Regression Model function}
\usage{
get.logistic.regression(dataset, indep, dep, training.index)
}
\arguments{
\item{dataset}{a data frame for data}

\item{indep}{a characters or a vector of characters for independent variables}

\item{dep}{a characters for a dependent variable}

\item{training.index}{a vector of numerics for training data indexes}
}
\description{
This function makes life simple by automatically fitting a logistic regression model.
}
\keyword{LR}
