% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.plot.data.stat.R
\name{get.plot.data.stat}
\alias{get.plot.data.stat}
\title{Get Statistical Values of Boxplots Data}
\usage{
get.plot.data.stat(input.plot.data, summary.index, file.name)
}
\arguments{
\item{input.plot.data}{a data frame for data}

\item{summary.index}{}
}
\description{
This function makes life simple by automatically summarising statistical values of boxplots data.
}
\keyword{summary}
