% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.predicted.probability.R
\name{get.predicted.probability}
\alias{get.predicted.probability}
\title{Get predicted probability}
\usage{
get.predicted.probability(model, test.data, indep)
}
\arguments{
\item{model}{a logistic regression or random forest model}

\item{test.data}{a data frame of testing dataset}

\item{indep}{a vector for independent variables}
}
\description{
This function makes life simple by automatically retrieving predicted probability of logistic regression and random forest.
}
\keyword{predict}
