% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.random.forest.classification.R
\name{get.random.forest.classification}
\alias{get.random.forest.classification}
\title{Get Random Forest model function (Classification)}
\usage{
get.random.forest.classification(
  dataset,
  indep,
  dep,
  training.index,
  ntree = 100
)
}
\arguments{
\item{dataset}{a data frame for data}

\item{indep}{a characters or a vector of characters for independent variables}

\item{dep}{a characters for a dependent variable}

\item{training.index}{a vector of numerics for training data indexes}

\item{ntree}{a numeric for the number of trees}
}
\description{
This function makes life simple by automatically fitting a logistic regression model.
}
\keyword{RF}
