% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.vc.R
\name{get.vc}
\alias{get.vc}
\title{Get VarClus based on the absolute Spearman correlation coefficients between metrics}
\usage{
get.vc(dataset, indep)
}
\arguments{
\item{dataset}{a data frame for data}

\item{indep}{a characters or a vector of characters for independent variables}
}
\description{
This function makes life simple by providing a VarClus.
}
\keyword{VarClus}
