% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImageURL.R
\name{getImageURL}
\alias{getImageURL}
\title{Get Image's URL by Google Ajax API}
\usage{
getImageURL(word, ...)
}
\arguments{
\item{Undescribed}{}
}
\value{
Undescribed
}
\description{
Get Image's URL by Google Ajax API
}
\examples{
# getImageURL(word = 'Paris Hilton', v = '1.0', rsz = 1, start = 1)
}

