% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_unite.R
\name{group_by_unite}
\alias{group_by_unite}
\title{Undescribe}
\usage{
group_by_unite(data, ..., col_n, sep = "_", remove = TRUE)
}
\arguments{
\item{data}{Undescribed}

\item{...}{Undescribed}

\item{col_n}{Undescribed}

\item{sep}{Undescribed}

\item{remove}{Undescribe}
}
\value{
Undescribe
}
\description{
Undescribe
}
\examples{
iris \%>\% group_by_unite(Petal.Width, Species)
iris \%>\% group_by_unite(Petal.Width, Species, col_n = 'unite_name', sep = '|')
}

