% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_plot.R
\name{hclust_plot}
\alias{hclust_plot}
\title{Custom dendrogram plot}
\usage{
hclust_plot(data, cut.num = 5, clust.var = TRUE, horiz = FALSE,
  labels_track_height = 120, main = "", ...)
}
\arguments{
\item{data}{Choose the dataframe.}

\item{cut.num}{Choose the number where to cut the number of clusters. Default: 5.}

\item{clust.var}{Choose if you want to cluster variables. Default: TRUE.}

\item{horiz}{Plot horizontal (TRUE) or vertical (FALSE). Default: FALSE.}
}
\description{
Make a custom dendrogram plot.
}
\examples{
hclust_plot(data, cut.num = 5)
}
\keyword{dataframe}
