% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_var.R
\name{hclust_var}
\alias{hclust_var}
\title{Custom dendrogram plot}
\usage{
hclust_var(data, cut.num, horiz = FALSE, ...)
}
\arguments{
\item{data}{Choose the dataframe.}

\item{cut.num}{Choose the number where to cut the output filename. Default: output/temp-results.txt}

\item{horiz}{Plot horizontal (TRUE) or vertical (FALSE). Default: FALSE.}
}
\description{
Make a custom dendrogram plot.
}
\examples{
hclust_var(data, cut.num = 5)
}
\keyword{dataframe}
