% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilmarinen_theme.R
\name{ilmarinen_cols}
\alias{ilmarinen_cols}
\title{Function to extract Ilmarinen colors as hex codes}
\usage{
ilmarinen_cols(...)
}
\arguments{
\item{...}{Character names of ilmarinen_colors}
}
\description{
Function to extract Ilmarinen colors as hex codes
}
