% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilmarinen_cols.R
\name{ilmarinen_pal}
\alias{ilmarinen_pal}
\title{Return function to interpolate Ilmarinen color palette}
\usage{
ilmarinen_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in ilmarinen_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate Ilmarinen color palette
}
