% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{inspect}
\usage{
inspect(data, filename = "./output/temp-results.txt", view = FALSE)
}
\arguments{
\item{data}{Choose the dataframe.}

\item{filename}{Choose the output filename. Default: ./output/temp-results.txt}

\item{view}{View the data inside R. Default: FALSE.}
}
\description{
Inspect the dataframe with str, names, summary, describe and View.
}
\examples{
inspect(data, "./output/temp-results.txt")
}
\keyword{dataframe}
\keyword{inspect}
