% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_date}
\alias{is_date}
\title{is_date}
\usage{
is_date(x)
}
\arguments{
\item{x}{Date type.}
}
\description{
Check if variable is a date or a datetime object.
}
\examples{
is_date(Sys.Date())
}
\keyword{bizday}
\keyword{date}
\keyword{weekday}
