% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_named_vector.R
\name{is_named_vector}
\alias{is_named_vector}
\title{is_named_vector}
\usage{
is_named_vector(x, mode = "any")
}
\arguments{
\item{x}{Choose vector to test.}

\item{mode}{character string naming an atomic mode or "list" or "expression"
or (except for vector) "any". Allows any type (see typeof) for mode,
and when mode is not "any" almost the same as typeof(x) == mode.}
}
\description{
Check if is named vector
}
\examples{
is_named_vector(c(letter = "a", number = 4))
}
\keyword{vector}
