% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmisc.R
\name{is_whole_number}
\alias{is_whole_number}
\title{A function to check whether a numeric vector contains no decimals.}
\usage{
is_whole_number(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A numeric vector}

\item{tol}{The tolerance allowed}
}
\value{
A single logical value
}
\description{
Check whether the numeric vector can be converted to type integer
without loosing information.
}
\examples{
is_whole_number(c(0.0, 1, NA))
}
\seealso{
\code{\link[=is.integer]{is.integer()}} which only checks the type
}
