% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsummary.R
\name{jsummary.data.frame}
\alias{jsummary.data.frame}
\title{Title}
\usage{
\method{jsummary}{data.frame}(object, maxsum = 8L, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{A dataframe}

\item{maxsum}{Maximum number of levels shown}

\item{digits}{Passed to signif()}

\item{...}{Additional arguments}
}
\value{
A table
}
\description{
Title
}
\examples{
jsummary(attenu)
}
