% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsummary.R
\name{jsummary.Date}
\alias{jsummary.Date}
\title{Title}
\usage{
\method{jsummary}{Date}(object, digits = 12L, ...)
}
\arguments{
\item{object}{A Date object}

\item{digits}{Number of significant digits for computations}

\item{...}{Additional arguments}
}
\value{
A summaryDefault object
}
\description{
Title
}
\examples{
jsummary(as.Date(0:10))
}
