% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsummary.R
\name{jsummary.difftime}
\alias{jsummary.difftime}
\title{Title}
\usage{
\method{jsummary}{difftime}(
  object,
  digits = getOption("digits"),
  show_quantiles = TRUE,
  sd = FALSE,
  ...
)
}
\arguments{
\item{object}{A difftime object}

\item{digits}{Number of significant digits for computations}

\item{show_quantiles}{A logical value}

\item{sd}{A logical value. Whether to show standard deviation}

\item{...}{Additional arguments}
}
\value{
A summaryDefault object
}
\description{
Title
}
\examples{
jsummary(Sys.Date() - as.Date("2026-01-01"))
}
