% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsummary.R
\name{jsummary.factor}
\alias{jsummary.factor}
\title{Title}
\usage{
\method{jsummary}{factor}(object, maxsum = 100L, ...)
}
\arguments{
\item{object}{A factor}

\item{maxsum}{Maximum number of levels to show}

\item{...}{Additional arguments}
}
\value{
A named vector
}
\description{
Title
}
\examples{
v <- factor(rep(c("a", "b"), c(3, 2)), levels=c("a", "b", "c"))
jsummary(v)
}
