% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsummary.R
\name{jsummary}
\alias{jsummary}
\title{Generic function to produce privacy aware object summaries}
\usage{
jsummary(object, ...)
}
\arguments{
\item{object}{An object to summarise}

\item{...}{Additional arguments}
}
\value{
Depends on the class of the argument
}
\description{
Generic function to produce privacy aware object summaries
}
\examples{
jsummary(1:10)
}
