% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_set_wrapper.R
\name{key_set_wrapper}
\alias{key_set_wrapper}
\title{key_set_wrapper}
\usage{
key_set_wrapper(service, username = NULL, keyring = NULL)
}
\arguments{
\item{service}{Choose keyring service.}

\item{username}{Choose optional username.}

\item{keyring}{Choose optional keyring for extra security.}
}
\description{
If else conditions for keyring::key_set
}
\examples{
library(keyring)
key_set_wrapper("test_keyring_20200101")
key_delete("test_keyring_20200101")
}
\keyword{keyring,}
\keyword{wrapper}
