% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmNatrisk.R
\name{kmNatrisk}
\alias{kmNatrisk}
\title{kmNatrisk}
\usage{
kmNatrisk(km, mark = 3, simple = FALSE, xaxis.at = pretty(km$time),
  xaxis.lab = xaxis.at, lty.surv = 1, lwd.surv = 1, col.surv = 1,
  lty.ci = 0, lwd.ci = 1, col.ci = col.surv, group.names = NULL,
  group.order = seq(length(km$n)), extra.left.margin = 4,
  label.n.at.risk = FALSE, draw.lines = TRUE, cex.axis = 1, xlab = "",
  ylab = "", main = "", xlim = c(0, max(km$time)), ylim = c(0, 1),
  grid = TRUE, lty.grid = 1, lwd.grid = 1, col.grid = grey(0.9),
  legend = !is.null(km$strata), loc.legend = "topright", add = FALSE, ...)
}
\arguments{
\item{km}{output from \code{survfit} function in \code{survival} package.}
}
\value{
draws Kaplan Meier curves with N at risk.
}
\description{
Survival curves displaying number at risk.
}
\author{
JG
}
