% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Linear model}
\usage{
linear_model(data, form = c("y ~ ."), k = 10)
}
\arguments{
\item{data}{Name of the data frame.}

\item{form}{Model in the format c("variable ~ ."). Default: Predict variable y with all other variables.}

\item{k}{Number of kfolds. Default: 10.}
}
\description{
Function for linear model and its results
}
\examples{
linear_model(data, c("y ~ x + z"))
}
\keyword{linear}
\keyword{model}
