% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_email.R
\name{list_to_email}
\alias{list_to_email}
\title{list_to_email}
\usage{
list_to_email(source = readClipboard())
}
\arguments{
\item{my.input.df}{Enter your data frame.}
}
\description{
This function helps you to save O365 type email addresses from list
}
\examples{
list_to_email()
}
\keyword{email}
\keyword{list}
