% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pkg.R
\name{load_pkg}
\alias{load_pkg}
\title{load_pkg}
\usage{
load_pkg(packages = "full")
}
\arguments{
\item{packages}{"full" loads the full list of favourite packages and "basic" loads tidyverse, data.table and a few other packages.}
}
\description{
This function installs and loads your favourite packages.
}
\examples{
load_pkg(packages = "long")
}
\keyword{packages}
