% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_model.R
\name{logit_model}
\alias{logit_model}
\title{Logit model}
\usage{
logit_model(data, form = c("y ~."), k = 10)
}
\arguments{
\item{data}{Name of the data frame.}

\item{form}{Model in the format c("variable ~ ."). Default: Predict variable y with all other variables.}

\item{k}{Number of kfolds. Default: 10.}
}
\description{
Function for logarithmic model and its results
}
\examples{
logit_model(data, c("y ~ x + z"))
}
\keyword{linear}
\keyword{model}
