% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreguni.R
\name{logreguni}
\alias{logreguni}
\title{logreguni}
\usage{
logreguni(x, y)
}
\arguments{
\item{x}{a data frame containing the independent variables.}

\item{y}{the response variable, typically 0 or 1.}
}
\value{
prints tables with coefficients and confidence intervals.
}
\description{
Calculate simple logistic regression models and return a formatted output.
}
\author{
Jutta Gamper
}
\keyword{htest}
