% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Return mode values}
\usage{
Mode(x, useNA = FALSE, ...)
}
\arguments{
\item{x}{A vector of values}

\item{useNA}{Use NA or not}

\item{...}{Arguments that table function have}
}
\value{
mode values
}
\description{
Return mode values
}
\examples{
dat <- sample(10, 100, TRUE)
Mode(dat)
dat_na <- sample(c(1, NA), 100, TRUE, c(0.1, 0.9))
Mode(dat_na, useNA = TRUE)
}

