% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot_bars.R
\name{multiplot_bars}
\alias{multiplot_bars}
\title{Multiple custom bar plots}
\usage{
multiplot_bars(data, row = 1:nrow(.), labels = waiver())
}
\arguments{
\item{data}{dataframe}

\item{row}{select which row to plot. Default: 1:nrow(.)}

\item{label}{select labels. Default: waiver()}
}
\description{
Create a multiple custom bar plots. With purr::map you can create a list of barplots.
}
\examples{
plot.basic <- data \%>\%
select(1:4) \%>\%
{map(list(1, 2, c(3:5)), function(x) multiplot_bars(., x, labels))}
do.call("grid.arrange", c(plot.basic, ncol = 2))

multiplot_bars(data, row = 3)
}
\keyword{ggplot2}
\keyword{multiplot}
\keyword{plot}
