% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{next_weekday}
\alias{next_weekday}
\title{next_weekday}
\usage{
next_weekday(x)
}
\arguments{
\item{x}{Date type.}
}
\description{
Choose the next weekday for a date
}
\examples{
next_weekday(Sys.Date())
}
\keyword{bizday}
\keyword{date}
\keyword{weekday}
