% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{notify}
\alias{notify}
\title{notify command by URL(http://uribo.hatenablog.com/entry/2015/09/22/120815)}
\usage{
notify(msg = "Operation complete")
}
\arguments{
\item{msg}{Undescribed}
}
\value{
Undescribe
}
\description{
notify command by URL(http://uribo.hatenablog.com/entry/2015/09/22/120815)
}

