% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\title{Path function}
\usage{
path(directory)
}
\arguments{
\item{directory}{a chatacters for input directory}
}
\description{
This function allows you to safely check if the input path is created. If not, then the input path will be automatically created. Finally, the function return the input path to lessen life complication.
}
\keyword{Path}
