% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R, R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\alias{format_num}
\title{format_num}
\usage{
format_num(x)

lhs \%>\% rhs
}
\arguments{
\item{x}{Choose the variable to round to 1 decimal with , as a decimal mark.}
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\examples{
x <- seq(0, 10, by = 0.01)
format_num(x)
}
\keyword{format,}
\keyword{internal}
\keyword{numeric}
