% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{plot_corr}
\alias{plot_corr}
\title{Custom correlation plot}
\usage{
plot_corr(data, polychoric = FALSE)
}
\arguments{
\item{data}{Choose the dataframe.}

\item{polychoric}{Not yet implemented. TRUE = use polychoric correlations.}
}
\description{
Make a custom correlation plot.
}
\examples{
plot_corr(mtcars)
}
\keyword{plot}
