% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing.R
\name{plot_missing}
\alias{plot_missing}
\title{plot_missing}
\usage{
plot_missing(data, file = "plot_missmap.png")
}
\arguments{
\item{data}{Enter your data frame.}
}
\description{
This function helps you to ggplot missing values of a data frame.
}
\examples{
ggplot_missing(data, "./output/plot_missmap.png")
}
\keyword{ggplot2}
\keyword{missing}
