% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psum.R
\name{psum}
\alias{psum}
\title{parallel sum}
\usage{
psum(..., na.rm = FALSE)
}
\arguments{
\item{...}{numeric or character arguments}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\description{
parallel sum
}

