% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questback_api.R
\name{qb_get_questions}
\alias{qb_get_questions}
\title{qb_get_questions}
\usage{
qb_get_questions(username, password, quest_id, security_id)
}
\arguments{
\item{username}{Username to Questback.}

\item{password}{Password to Questback.}

\item{quest_id}{Quest id.}

\item{security_id}{Security id.}
}
\description{
Get responses to a survey from Questback Essentials.
}
\keyword{Questback,}
\keyword{api}
\keyword{survey,}
