% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questback_api.R
\name{qb_get_quests}
\alias{qb_get_quests}
\title{qb_get_quests}
\usage{
qb_get_quests(username, password, n = 1000)
}
\arguments{
\item{username}{Username to Questback.}

\item{password}{Password to Questback.}

\item{n}{Maximum number of results. Max: 1000.}
}
\description{
Get responses to a survey from Questback Essentials.
}
\keyword{Questback,}
\keyword{api}
\keyword{survey,}
