% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questback_integration.R
\name{qb_send_invitees}
\alias{qb_send_invitees}
\title{qb_send_invitees}
\usage{
qb_send_invitees(
  data,
  quest_id,
  sid,
  username,
  password,
  filename = tempfile(),
  transfer_dir = NULL,
  email = "email",
  allow_duplicate = FALSE
)
}
\arguments{
\item{data}{Data to import the invitees from.}

\item{quest_id}{Quest id from Questback Essentials.}

\item{sid}{Quest security id from Questback Essentials.}

\item{username}{Username for Questback Essentials.}

\item{password}{Password for Questback Essentials.}

\item{filename}{Filename to save the data temporarely.}

\item{transfer_dir}{Directory where to saave the imported data.}

\item{email}{Name of the email field.}

\item{allow_duplicate}{Allow to import duplicate email fields.}
}
\description{
Send invitees to a survey from Questback Essentials.
}
\keyword{essentials,}
\keyword{questback}
\keyword{survey}
