% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questback_api.R
\name{qb_test_connection}
\alias{qb_test_connection}
\title{qb_test_connection}
\usage{
qb_test_connection(username, password)
}
\arguments{
\item{username}{Username to Questback.}

\item{password}{Password to Questback.}
}
\description{
Test connection to Questback.
}
\keyword{Questback,}
\keyword{api}
\keyword{survey,}
