% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_clip}
\alias{rcpp_clip}
\title{Clip vector between numbers a and b}
\usage{
rcpp_clip(x, a, b)
}
\arguments{
\item{x}{Input vector}

\item{a}{Minumum value}

\item{b}{Maximum value}
}
\description{
Clip vector between numbers a and b
}
