\name{repCol}
\alias{repCol}
\title{Repeat a vector by col}
\usage{
repCol(x, n)
}
\arguments{
  \item{x}{\code{vector} or \code{matrix}}

  \item{n}{number of replication}
}
\description{
Repeat a vector by col
}
\examples{
repRow(c(a=1,b=2,c=3),5)
repCol(c(a=1,b=2,c=3),5)
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}
\seealso{
\link{repRow}
}

