% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaces.R
\name{replaces}
\alias{replaces}
\title{Undescribe}
\usage{
replaces(x, from, to, last = FALSE, evaluate = TRUE)
}
\arguments{
\item{Undescribe}{}
}
\value{
Undescribe
}
\description{
Undescribe
}
\examples{
replaces(1:10, 1:3, 3:5)
}

