\name{repRow}
\alias{repRow}
\title{Repeat a vector by row}
\usage{
repRow(x, n)
}
\arguments{
  \item{x}{\code{vector} or \code{matrix}}

  \item{n}{number of replication}
}
\description{
Repeat a vector by row
}
\examples{
repRow(c(a=1,b=2,c=3),5)
repCol(c(a=1,b=2,c=3),5)
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}
\seealso{
\link{repCol}
}

