% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/require_force.R
\name{require_force}
\alias{require_force}
\title{Undescribed}
\usage{
require_force(package, ..., load = TRUE)
}
\arguments{
\item{Undescribed}{}
}
\value{
Undescribed
}
\description{
Undescribed
}
\examples{
require_force(dplyr)
}

