% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rGEO.R
\name{rGEO}
\alias{rGEO}
\title{Reverse Geocoding by Google Maps API}
\usage{
rGEO(latlng, para = list(sensor = "false", language = "ja", region = "JP"),
  proxy = NULL)
}
\arguments{
\item{Undescribed}{}
}
\value{
Undescribed
}
\description{
Reverse Geocoding by Google Maps API
}

