% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scale2.R
\name{scale2}
\alias{scale2}
\title{This let x's scale correspond y's scale.}
\usage{
scale2(x, y)
}
\arguments{
\item{x}{A vector of values}

\item{y}{A vector of values}
}
\value{
scaled vector \code{x}
}
\description{
This let x's scale correspond y's scale.
}
\examples{
scale2(1:10, 0:1)
}

