% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_error_mail.R
\name{send_error_mail}
\alias{send_error_mail}
\title{send_error_mail}
\usage{
send_error_mail(to, subject = "", body = "", user, password, smtp = NULL)
}
\arguments{
\item{to}{email or a vector of emails.}

\item{subject}{of an email.}

\item{body}{of en email.}

\item{user}{user of sender's email account}

\item{password}{password of sender's email account}

\item{smtp}{smtp. If NULL, then gmail.}
}
\description{
Send error email.
}
\examples{
\dontrun{
to <- "address@email.com"
subject <- "You have error"
body <- "Hi,
you received an error. Check your logs."
user <- "erroraddress@email.com"
password <- "1234567"
send_error_mail(to, subject, body, user, password)
stop("Error")
}
}
\keyword{email,}
\keyword{error}
