% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_error_sms.R
\name{send_error_sms}
\alias{send_error_sms}
\title{send_error_sms}
\usage{
send_error_sms(to, twilio_phone)
}
\arguments{
\item{to}{phone number or a vector of phone numbers to send the error message.}

\item{twilio_phone}{Twilio phone number from your twilio account.}
}
\description{
Send SMS on error.
}
\examples{
\dontrun{
to <- "0401112222"
twilio_phone <- "04011113333"
send_error_sms(to, twilio_phone)
stop("Error")
}
}
\keyword{SMS,}
\keyword{error}
