% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_error_toast.R
\name{send_error_toast}
\alias{send_error_toast}
\title{send_error_toast}
\usage{
send_error_toast(text = "This is a toast with error:", title = "Alert!")
}
\arguments{
\item{text}{Text of the toast. Default: This is a toast}

\item{title}{Title of the toast. Default: Alert!}
}
\description{
Send error toast.
}
\examples{
\dontrun{
send_error_toast("Something went wrong")
stop("Error")
}
}
\keyword{error}
\keyword{toast,}
