% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_mail.R
\name{send_mail}
\alias{send_mail}
\title{send_mail}
\usage{
send_mail(
  to,
  subject = "",
  body = "",
  attachment = NULL,
  signature = TRUE,
  encoding = "UTF-8"
)
}
\arguments{
\item{to}{email or a vector of emails.}

\item{subject}{of an email.}

\item{body}{of en email.}

\item{attachment}{path to attachment file. Default: NULL}

\item{signature}{adds signature taken from Outlook. Default: TRUE}

\item{encoding}{encoding of the body. Default: "UTF-8"}
}
\description{
Send email from Outlook.
}
\examples{
\dontrun{
to <- c("jouni.vatanen@ilmarinen.fi", "jouni.vatanen@ilmarinen.fi")
subject <- "Ääkkösiä"
body <- "Hei,
meillä on fiksuja asioita
ääkkösiä
vai miten se oli"
send_mail(to, subject, body)
}
}
\keyword{email,}
\keyword{outlook}
