% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_error_toast.R
\name{send_toast}
\alias{send_toast}
\title{send_toast}
\usage{
send_toast(text, title)
}
\arguments{
\item{text}{Text of the toast.}

\item{title}{Title of the toast.}
}
\description{
Send toast.
}
\examples{
\dontrun{
send_toast("You got toast")
}
}
\keyword{toast}
