% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_rownames.R
\name{set_rownames}
\alias{set_rownames}
\title{Undescribed}
\usage{
set_rownames(df, var, sep = ".")
}
\arguments{
\item{Undescribed}{}
}
\value{
Undescribed
}
\description{
Undescribed
}
\examples{
iris \%>\% set_rownames(Species) \%>\% head
iris \%>\% set_rownames('Species') \%>\% head
}

