% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setProxy.R
\name{setProxy}
\alias{setProxy}
\title{Undescribed}
\usage{
setProxy(url = Sys.getenv("PROXY_URL"), port = Sys.getenv("PROXY_PORT"))
}
\arguments{
\item{Undescribed}{}
}
\value{
Undescribed
}
\description{
Undescribed
}

