% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift.R
\name{shift}
\alias{shift}
\title{Shift vector's elements}
\usage{
shift(x, n = 1L, back = TRUE)
}
\arguments{
\item{x}{A vector of values}

\item{n}{A integer of length 1. x will shift by the number.}

\item{back}{If TRUE, will shift \code{x} back by \code{n}.}
}
\value{
shifted \code{x}
}
\description{
Shift vector's elements
}
\examples{
shift(1:10)
shift(1:10, n = 3L)
shift(1:10, back = FALSE)
shift(1:10, n = -1L)
}

