\name{sortCol}
\alias{sortCol}
\title{Sort the value of column for each row of a matrix}
\usage{
sortCol(x, decreasing = FALSE)
}
\arguments{
  \item{x}{a matrix}

  \item{decreasing}{in descending order. Default is FALSE}
}
\description{
Sort the value of
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

