% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveypal_api.R
\name{sp_fromJSON}
\alias{sp_fromJSON}
\title{sp_fromJSON}
\usage{
sp_fromJSON(json)
}
\arguments{
\item{json}{JSON type value.}
}
\description{
Parse JSON from Surveypal to dataframe.
}
\keyword{api}
\keyword{survey,}
\keyword{surveypal,}
