% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveypal_api.R
\name{sp_get_answers}
\alias{sp_get_answers}
\title{sp_get_answers}
\usage{
sp_get_answers(
  survey_id,
  api_key,
  locale = NULL,
  answer_state = NULL,
  from = NULL,
  to = NULL,
  limit = NULL,
  use_question_id = FALSE,
  encoding = "UTF-8"
)
}
\arguments{
\item{survey_id}{Quest id from Surveypal.}

\item{api_key}{Api-key from Surveypal.}

\item{locale}{Locale. Default: first locale in the survey.}

\item{answer_state}{Choises are "[all]", "[complete]", "[incomplete]", "[unstarted]",}

\item{from}{First datetime to get the data. In the form YYYY-MM-DD,
YYYY-MM-DD HH:mm:ss or DD.MM.YYYY.}

\item{to}{Last datetime to get the data. In the form YYYY-MM-DD,
YYYY-MM-DD HH:mm:ss or DD.MM.YYYY.}

\item{limit}{Limit the number of returned answers.}

\item{use_question_id}{Show element question id instead of question text.
Default: false.}
}
\description{
Get answers to a survey from Surveypals.
}
\keyword{api}
\keyword{survey,}
\keyword{surveypal,}
