% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveypal_api.R
\name{sp_get_elements}
\alias{sp_get_elements}
\title{sp_get_elements}
\usage{
sp_get_elements(
  survey_id,
  api_key,
  locale = NULL,
  use_question_id = FALSE,
  encoding = "UTF-8"
)
}
\arguments{
\item{survey_id}{Survey id from Surveypal.}

\item{api_key}{Api-token from Surveypal.}

\item{locale}{Locale. Default: First locale of the survey. Example value: "fi_FI".}

\item{use_question_id}{Use question id. Default is FALSE.}
}
\description{
Get elements from a single Surveypal survey.
}
\keyword{api}
\keyword{survey,}
\keyword{surveypal,}
