% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveypal_api.R
\name{sp_get_responses}
\alias{sp_get_responses}
\title{sp_get_responses}
\usage{
sp_get_responses(
  filename,
  quest_id,
  sid,
  username,
  password,
  last_updated = "12 hours",
  latest_days = NULL,
  from_date = NULL,
  to_date = NULL
)
}
\arguments{
\item{filename}{Filename to save the imported data.}

\item{last_updated}{If time is less than chosen, then do not execute.
Possible time choises are secs, mins, hours, days and weeks. Default is 12 hours.}

\item{latest_days}{From how many days you get the data.}

\item{from_date}{First datetime to get the data. In the form YYYY-MM-DD or
YYYY-MM-DD HH:mm:ss}

\item{to_date}{Last datetime to get the data. In the form YYYY-MM-DD or
YYYY-MM-DD HH:mm:ss}

\item{survey_id}{Quest id from Surveypal.}

\item{api_token}{Api-token from Surveypal.}
}
\description{
Get responses to a survey from Surveypals.
}
\keyword{api}
\keyword{survey,}
\keyword{surveypal,}
