% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveypal_api.R
\name{sp_get_surveys}
\alias{sp_get_surveys}
\title{sp_get_surveys}
\usage{
sp_get_surveys(
  api_key,
  published = NULL,
  from = NULL,
  to = NULL,
  limit = NULL,
  offset = NULL,
  encoding = "UTF-8"
)
}
\arguments{
\item{api_key}{Api-key from Surveypal.}

\item{published}{Filter survey list by survey published status: TRUE/FALSE}

\item{from}{Include survey if modified after date.}

\item{to}{Include survey if modified before date.}

\item{limit}{The maximum number of surveys to return.}

\item{offset}{The offset of first survey to return.}
}
\description{
Get surveys from Surveypal.
}
\keyword{api}
\keyword{survey,}
\keyword{surveypal,}
