% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveypal_api.R
\name{sp_parse_datetime}
\alias{sp_parse_datetime}
\title{sp_parse_datetime}
\usage{
sp_parse_datetime(x)
}
\arguments{
\item{x}{Date or datetime object or an object convertable to datetime object.}
}
\description{
Transform datetime object to the format Surveypal recognizes. Or tries to
convert a character to correct format.
}
\keyword{api}
\keyword{survey,}
\keyword{surveypal,}
