% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveypal_api.R
\name{sp_send_invitees}
\alias{sp_send_invitees}
\title{sp_send_invitees}
\usage{
sp_send_invitees(
  data,
  survey_id,
  api_token,
  email = "email",
  unique = FALSE,
  filename = NULL,
  anonymous = FALSE,
  subject,
  message,
  language,
  from,
  reply_to,
  send_at
)
}
\arguments{
\item{data}{Data to import the invitees from.}

\item{survey_id}{Survey id from Surveypal.}

\item{api_token}{Api-token from Surveypal.}

\item{email}{Name of the email field.}

\item{unique}{Is the email unique.}

\item{filename}{Filename to save the sent list.}

\item{anonymous}{Is the answer anonymous.}

\item{subject}{Email subject.}

\item{message}{Email message.}

\item{language}{Language of the survey.}

\item{from}{From email address.}

\item{reply_to}{Reply to email address.}

\item{send_at}{When to send the survey emails.}
}
\description{
Send invitees to a survey from Surveypal.
}
\keyword{api}
\keyword{survey,}
\keyword{surveypal,}
