% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveypal_api.R
\name{sp_send_invites}
\alias{sp_send_invites}
\title{sp_send_invites}
\usage{
sp_send_invites(
  json,
  survey_id,
  api_key,
  anonymous = FALSE,
  unique_emails = TRUE,
  subject = NULL,
  message = NULL,
  locale = NULL,
  author = NULL,
  reply_to = NULL,
  send_at = NULL
)
}
\arguments{
\item{json}{JSON to import the invitees from.}

\item{survey_id}{Survey id from Surveypal.}

\item{api_key}{Api-key from Surveypal.}

\item{anonymous}{Is the answer anonymous. Default: FALSE.}

\item{unique_emails}{Is the email unique. Default: TRUE.}

\item{subject}{Email subject. NULL uses subject defined at the Surveypal.}

\item{message}{Email message. NULL uses message defined at the Surveypal.}

\item{author}{The invitation sender name.}

\item{reply_to}{The invitation reply to email address.}

\item{send_at}{The date and time when to send invitations.}

\item{language}{Language of the survey. Default: first language of the survey.}
}
\description{
Send invites to a survey from Surveypal.
}
\keyword{api}
\keyword{survey,}
\keyword{surveypal,}
