% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveypal_api.R
\name{sp_test_connection}
\alias{sp_test_connection}
\title{sp_test_connection}
\usage{
sp_test_connection(api_key)
}
\arguments{
\item{api_key}{Api-key from Surveypal.}
}
\description{
Test connection to Surveypal.
}
\keyword{api}
\keyword{survey,}
\keyword{surveypal,}
