% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveypal_api.R
\name{sp_toJSON}
\alias{sp_toJSON}
\title{sp_toJSON}
\usage{
sp_toJSON(.data, email)
}
\arguments{
\item{.data}{Data to parse to JSON.}

\item{email}{Name of the email field.}
}
\description{
Parse data to JSON to Surveypal format.
}
\keyword{api}
\keyword{survey,}
\keyword{surveypal,}
