% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise.vif.R
\name{stepwise.vif}
\alias{stepwise.vif}
\title{Automated Stepwise Variance Inflation Factor Analysis}
\usage{
stepwise.vif(dataset, indep, vif.threshold = 5, verbose = F, export = F,
  export.path = ".", export.file.name = "vif.results.csv")
}
\arguments{
\item{dataset}{a data frame for data}

\item{indep}{a characters or a vector of characters for independent variables}

\item{vif.threshold}{a numeric for a threshold of VIF score (default = 5)}

\item{verbose}{TRUE for printing}

\item{export}{TRUE for exporting results of the analysis}

\item{export.path}{a path for export file}

\item{export.file.name}{a file name for export file}
}
\description{
This function makes life simple by automatically excluding the most predictable metric according to VIF in a stepwise manner. For each step, the most predictable metric that has its VIF score of above the pre-defined threshold (default = 5) will be excluded. Such process is repeated untill all remaining metrics have their VIF score below the threshold.
}
\keyword{VIF}
