% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.grouped_unite.R
\name{summarise}
\alias{summarise}
\title{Undescribe}
\usage{
summarise(.data, ...)
}
\arguments{
\item{.data}{Undescribed}

\item{...}{Undescribed}
}
\value{
Undescribe
}
\description{
Undescribe
}

