% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumNum.R
\name{sumNum}
\alias{sumNum}
\title{sumNum}
\usage{
sumNum(data, grp = NULL, rd = 2)
}
\arguments{
\item{data}{a data frame or matrix.}

\item{grp}{possible factor to group by, defaults to NULL. Maximum are two groups.}

\item{rd}{round results to \code{rd} digits, default is 2.}
}
\value{
prints a summary table of number of observations (N), number of missing values (miss), minimum, maximum, median,
mean and standard deviation; possibly separated by grp.
}
\description{
Summarise all numerical variables in a \code{data.frame}.
}
\examples{

sumNum(iris)
sumNum(iris, 5)
sumNum(iris, 5, 4)
}
\author{
Jutta Gamper
}
