% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest}
\alias{ttest}
\title{ttest}
\usage{
ttest(x, y)
}
\arguments{
\item{x}{a data frame.}

\item{y}{the response variable.}
}
\value{
prints tables with means, standard deviations and p-values.
}
\description{
Calculate t-tests and return a table containing estimates and p-values.
}
\author{
Jutta Gamper
}
