% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfactor.R
\name{unfactor}
\alias{unfactor}
\title{Undescribe}
\usage{
unfactor(x, ...)
}
\arguments{
\item{x}{Undescribed}

\item{...}{Undescribed}
}
\value{
Undescribe
}
\description{
Undescribe
}
\examples{
unfactor(iris$Species)
}

