% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ununite.R
\name{ununite}
\alias{ununite}
\title{Undescribe}
\usage{
ununite(data, ...)
}
\arguments{
\item{data}{Undescribed}

\item{...}{Undescribed}
}
\value{
Undescribe
}
\description{
Undescribe
}
\examples{
iris \%>\% 
  group_by_unite(Petal.Width, Species) \%>\% 
  summarise(count = n()) \%>\% 
  ununite
iris \%>\% 
  group_by_unite(Petal.Width, Species, col_n = 'unite_name', sep = '|') \%>\% 
 summarise(count = n()) \%>\% 
 ununite
}

