% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.output.R
\name{write.output}
\alias{write.output}
\title{Return only first mode value}
\usage{
write.output(..., file = NULL, append = TRUE, cout = "## ")
}
\arguments{
\item{sequence}{of values}
}
\value{
first mode value
}
\description{
Return only first mode value
}
\examples{
write.output(head(CO2))
write.output(head(CO2), cout = '#out: ')
}

