% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csv_fi.R
\name{write_csv_fi}
\alias{write_csv_fi}
\title{write_csv_fi}
\usage{
write_csv_fi(
  x,
  path = ".",
  na = "",
  delim = "\\t",
  quote_escape = "double",
  append = FALSE,
  col_names = !append
)
}
\arguments{
\item{x}{Any list like element e.g. data.frame and data.table.}

\item{path}{Path or connection to write to.}

\item{na}{String used for missing values. Defaults to "". Missing values
will never be quoted; strings with the same value as na will always be quoted.}

\item{delim}{SDelimiter used to separate values. Defaults to "tab"}

\item{quote_escape}{The type of escaping to use for quoted values, one of
"double", "backslash" or "none". Defaults to "double".}

\item{append}{If FALSE, will overwrite existing file. If TRUE, will append to
existing file. In both cases, if file does not exist a new file is created.}

\item{col_names}{Write columns names at the top of the file? Must be either
TRUE or FALSE.}
}
\description{
Save any list like data.frame, data.table or matrix to a txt-file.
Uses readr::write_excel_csv2.
}
\examples{
n <- c(1.1, 2.2, 3.3)
s <- c("a", "b", "c")
x <- data.frame(n, s)
write_csv_fi(x, file = "example.txt")
}
\keyword{save}
\keyword{txt}
