% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_row2file.R
\name{write_row2file}
\alias{write_row2file}
\title{write_row2file}
\usage{
write_row2file(data, filename, na = "-")
}
\arguments{
\item{data}{Dataframe which to save.}

\item{filename}{Output filename.}

\item{na}{NA value.}
}
\description{
Write a row to a file
}
\examples{
write_row2file(mtcars[1, 1:2])
}
\keyword{clean,}
\keyword{company,}
\keyword{name}
\keyword{string,}
