% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xml_path2list}
\alias{xml_path2list}
\title{xml_path2list}
\usage{
xml_path2list(x, xpath)
}
\arguments{
\item{x}{XML file.}

\item{xpath}{Path of the XML file.}
}
\description{
Transforms xml xpath to list. Is used especially with Questback api functions.
xml2 functions have some kind of problems with nested named XML lists.
}
\keyword{Questback}
\keyword{XML}
