% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{anonymize}
\alias{anonymize}
\title{Anonymize}
\usage{
anonymize(x, algo = "crc32")
}
\arguments{
\item{algo}{Algorithm to anonymize id. Default: crc32.}

\item{data}{Choose the dataframe?.}
}
\description{
NOT FINISHED!!! Create anonymized id .
}
\examples{
anonymize(x, algo = "crc32")
}
\keyword{anonymize}
\keyword{data}
