% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_essentials.R
\name{api_essentials}
\alias{api_essentials}
\title{api_essentials}
\usage{
api_essentials(
  soap_action,
  username,
  password,
  quest_id = NULL,
  security_lock = NULL,
  x = NULL,
  send_duplicate = TRUE,
  sep = ";"
)
}
\arguments{
\item{soap_action}{API action: TestConnection / AddRespondentsData / AddEmailInvitees.}

\item{username}{Username for Questback Essentials.}

\item{password}{Password for Questback Essentials integration.}

\item{quest_id}{Quest id from Questback Essentials.}

\item{security_lock}{Security lock from Questback Essentials.}

\item{x}{the file name to save the dataframe.}

\item{send_duplicate}{Is it allowed to send survey to the same respondent.}

\item{sep}{Choose separator. Default: ";"}
}
\description{
API to POST/GET Questback Essentials survey information
}
\keyword{API}
\keyword{survey}
