% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_birth.R
\name{calc_birth}
\alias{calc_birth}
\title{calc_birth}
\usage{
calc_birth(x)
}
\arguments{
\item{x}{Choose string from which to calculate.}
}
\description{
Calculate birth date from Finnish social security number and return POSIXct
}
\examples{
calc_birth("121256-123S")
}
\keyword{birth,}
\keyword{date}
\keyword{security,}
\keyword{social}
