% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.R
\name{col}
\alias{col}
\title{Easy color linear interpolation}
\usage{
col(x, colors = c("blue", "red"))
}
\arguments{
\item{x}{A vector (or array) of values corresponded to the colors}

\item{colors}{A character vector of colors to interpolate}
}
\value{
The interpolated color code
}
\description{
Easy color linear interpolation
}
\examples{
x <- 1:10
plot(x, pch = 19, col = col(x))
plot(x, pch = 19, col = col(-x))
}

