% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/col.R
\name{col}
\alias{col}
\title{Undescribed}
\usage{
col(x, colors = c("blue", "red"))
}
\arguments{
\item{Undescribed}{}
}
\value{
Undescribed
}
\description{
Undescribed
}
\examples{
x <- 1:10
plot(x, pch = 19, col = col(x))
plot(x, pch = 19, col = col(-x))
}

