% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_plot.R
\name{corr_plot}
\alias{corr_plot}
\title{Custom correlation plot}
\usage{
corr_plot(data, polychoric = FALSE)
}
\arguments{
\item{data}{Choose the dataframe.}

\item{polychoric}{Not yet implemented. TRUE = use polychoric correlations.}
}
\description{
Make a custom correlation plot.
}
\examples{
corr_plot(data, polychoric = FALSE)
}
\keyword{plot}
