% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cortab.R
\name{cortab}
\alias{cortab}
\title{cortab}
\usage{
cortab(x, y)
}
\arguments{
\item{x}{a data frame.}

\item{y}{the response variable.}
}
\value{
prints tables with correlation coefficients and p-values.
}
\description{
Calculate Pearson correlation coefficients and return a table containing estimates and p-values.
}
\author{
Jutta Gamper
}
