% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_query.R
\name{db_get_query}
\alias{db_get_query}
\title{db_get_query}
\usage{
db_get_query(sql, drv = odbc::odbc(), ..., params = NULL)
}
\arguments{
\item{sql}{Sql query or sql file}

\item{drv}{Driver for database. Default odbc::odbc().}

\item{...}{Connection parameters like dsn or server, username and password
also extra parameters like encoding.}

\item{params}{List of parameters to replace question marks in sql query.
Default NULL}
}
\description{
Get query results and close connection safely
}
\keyword{DBI,}
\keyword{database,}
\keyword{dbGetQuery}
\keyword{odbc,}
