% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stat.R
\name{desc_stat}
\alias{desc_stat}
\title{Create descriptive statistics table}
\usage{
desc_stat(
  data,
  .select,
  group_by_cols = NULL,
  with_idk = 99,
  .labels = c(total = "N")
)
}
\arguments{
\item{data}{Choose dataframe.}

\item{.select}{Choose which columns you want to calculate distribution.}

\item{group_by_cols}{Group by chosen columns. Defaul: NULL}

\item{with_idk}{Data includes I don't know values. You want to compute them in distributions but not in means. Default: 99.}

\item{.labels}{Choose labels for table. Default: total = "N"}
}
\description{
Create descriptive statistics table
}
\examples{
desc_stat(mtcars, .select = c(2, 8:11), group_by_cols = "cyl")
}
\keyword{birth,}
\keyword{date}
\keyword{security,}
\keyword{social}
