% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_txt.R
\name{df_to_txt}
\alias{df_to_txt}
\title{df_to_txt}
\usage{
df_to_txt(data, file = "temp.txt", overwrite = FALSE)
}
\arguments{
\item{data}{Choose the dataframe.}

\item{overwrite}{Do you want to overwrite}

\item{filename}{Choose the file name to save the dataframe.}
}
\description{
Saves a dataframe to a txt-file.
}
\examples{
df_to_txt(data, filename = "example.txt")
}
\keyword{save}
\keyword{txt}
