% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_txt.R
\name{df_to_txt}
\alias{df_to_txt}
\title{df_to_txt}
\usage{
df_to_txt(x, file = "", sep = "\\t", dec = ",", overwrite = FALSE)
}
\arguments{
\item{x}{Any list like element e.g. data.frame and data.table.}

\item{file}{Output file name. Default: to the console.}

\item{sep}{Separator. Default: TAB}

\item{dec}{Decimal limiter. Default: ","}

\item{overwrite}{Overwrites the file, it it exists. Default: FALSE.}

\item{...}{Pass other parameters to fwrite like na, append, col.names}
}
\description{
Save any list like data.frame, data.table or matrix to a txt-file. Based on fwrite from data.table package.
}
\examples{
n <- c(1.1, 2.2, 3.3)
s <- c("a", "b", "c")
x <- data.frame(n, s)
df_to_txt(x, file = "example.txt")
}
\keyword{save}
\keyword{txt}
