% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fi_industries_2008}
\alias{fi_industries_2008}
\title{Finnish industry classification TOL2008}
\format{
A data frame with 89 rows and 5 variables:
\describe{
  \item{tol_0}{Combined some upper level classes (A-X)}
  \item{tol_1}{Upper level classes (A-X)}
  \item{tol_1_name}{Upper level classes names}
  \item{tol_2}{Lower level classes (00-99)}
  \item{tol_2_name}{Lower level classes names}
}
}
\usage{
fi_industries_2008
}
\description{
Data from Tilastokeskus about Finnish industry classification.
Data is from 2008. Data can be downloaded from:
https://www.stat.fi/meta/luokitukset/toimiala/versio.html
}
\keyword{datasets}
