% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{format_num}
\alias{format_num}
\title{format_num}
\usage{
format_num(
  x,
  digits = 1,
  format = "f",
  dec = ",",
  flag = "",
  big.mark = " ",
  ...
)
}
\arguments{
\item{x}{an atomic numerical or character object, typically a vector of
real numbers. Any class is discarded, with a warning.}

\item{digits}{the desired number of digits after the decimal point
(format = "f") or significant digits (format = "g", = "e" or = "fg"). If
digits is negative, then round to numbers like tens, hundreds or thousands.}

\item{dec}{he character to be used to indicate the numeric decimal point.}

\item{flag}{for formatC, a character string giving a format modifier as
in Kernighan and Ritchie (1988, page 243) or the C+99 standard. For example
"0" pads leading zeros; "-" does left adjustment;
"+" ensures a sign in all cases, i.e., "+" for positive numbers}

\item{big.mark}{used as mark between every big.interval decimals before
the decimal point.}

\item{...}{add other parameters to formatC like width and small.mark}
}
\description{
Formats numbers to the finnish standard.
}
\examples{
x <- seq(0, 10, by = 0.01)
format_num(x)
}
\keyword{format,}
\keyword{formatC,}
\keyword{numeric,}
\keyword{round}
