% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{GEO}
\alias{GEO}
\title{Geocoding by Google Maps API}
\usage{
GEO(address, para = list(sensor = "false", language = "ja", region = "JP"),
  proxy = NULL)
}
\arguments{
\item{Undescribed}{}
}
\value{
Undescribed
}
\description{
Geocoding by Google Maps API
}

