% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_missing.R
\name{ggplot_missing}
\alias{ggplot_missing}
\title{ggplot_missing}
\usage{
ggplot_missing(data, file = "plot_missmap.png")
}
\arguments{
\item{data}{Enter your data frame.}
}
\description{
This function helps you to ggplot missing values of a data frame.
}
\examples{
ggplot_missing(data, "output/plot_missmap.png")
}
\keyword{ggplot2}
\keyword{missing}
