% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_text.R
\name{image_text}
\alias{image_text}
\title{Undescribed}
\usage{
image_text(x, x_labels = TRUE, y_labels = TRUE, in_labels = FALSE, ...,
  mcex = 1L, tcex = 1L)
}
\arguments{
\item{x}{Undescribed}

\item{x_labels}{Undescribed}

\item{y_labels}{Undescribed}

\item{in_labels}{Undescribed}

\item{...}{Undescribed}

\item{mcex}{Undescribed}

\item{tcex}{Undescribed}
}
\value{
Undescribed
}
\description{
Undescribed
}
\examples{
image_text(iris[1:10, 1:4])
image_text(iris[1:10, 1:4], in_labels = TRUE)
}

