% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmisc-package.R
\docType{package}
\name{jmisc-package}
\alias{jmisc}
\alias{jmisc-package}
\title{jmisc: A set of miscellaneous utilities}
\description{
The package provides miscellaneous utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jttoivon/jmisc}
  \item Report bugs at \url{https://github.com/jttoivon/jmisc/issues}
}

}
\author{
\strong{Maintainer}: Jarkko Toivonen \email{jarkko.toivonen.19@gmail.com}

Authors:
\itemize{
  \item Jarkko Toivonen \email{jarkko.toivonen.19@gmail.com}
}

}
\keyword{internal}
