% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsummary.R
\name{jsummary.default}
\alias{jsummary.default}
\title{Privacy aware object summaries}
\usage{
\method{jsummary}{default}(
  object,
  ...,
  digits,
  quantile.type = 7,
  show_quantiles = TRUE,
  sd = TRUE
)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments}

\item{digits}{Passed to signif()}

\item{quantile.type}{Integer}

\item{show_quantiles}{A logical value}

\item{sd}{A logical value. Whether to show standard deviation}
}
\value{
Depends on the class of the argument
}
\description{
Privacy aware object summaries
}
\examples{
jsummary(1:10)
}
