% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsummary.R
\name{jsummary.POSIXct}
\alias{jsummary.POSIXct}
\title{Title}
\usage{
\method{jsummary}{POSIXct}(object, digits = 15L, show_quantiles = TRUE, sd = FALSE, ...)
}
\arguments{
\item{object}{A calendar date}

\item{digits}{Number of digits for fractional seconds}

\item{show_quantiles}{A logical value}

\item{sd}{A logical value. Whether to show standard deviation}

\item{...}{Additional arguments}
}
\value{
A summaryDefault object
}
\description{
Title
}
\examples{
jsummary(as.POSIXct(0:10))
}
