% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_email.R
\name{list_to_email}
\alias{list_to_email}
\title{list_to_email}
\usage{
list_to_email(source = readClipboard(), sep = "\\t", reverse = FALSE)
}
\arguments{
\item{source}{From where to read the data. Default: clipboard.}

\item{sep}{Separator, if data is not from clipboard. Default: tab.}

\item{reverse}{Reverse the order of the first and the last name. Default: FALSE.}
}
\description{
This function helps you to save O365 type email addresses from list
}
\examples{
list_to_email()
}
\keyword{email}
\keyword{list}
