% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pkg.R
\name{load_pkg}
\alias{load_pkg}
\title{load_pkg}
\usage{
load_pkg(packages = "long")
}
\description{
This function installs and loads your favourite packages.
}
\examples{
load_pkg(packages = "long")
}
\keyword{packages}
