% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW.R
\name{MW}
\alias{MW}
\title{Moving window}
\usage{
MW(dat, FUN, int = 2L)
}
\arguments{
\item{dat}{Sequence of values}

\item{FUN}{The function to be applied to each window}

\item{int}{Window size}
}
\value{
first mode value
}
\description{
Moving window
}
\examples{
MW(1:10, mean)
}

