% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_people_names.R, R/format_num.R
\name{\%>\%}
\alias{\%>\%}
\alias{clean_people_names}
\alias{format_num}
\title{clean_people_names}
\usage{
clean_people_names(x)

format_num(x)
}
\arguments{
\item{x}{Choose string which to clean.}

\item{x}{Choose the variable to round to 1 decimal with , as a decimal mark.}
}
\description{
Clean string from Finnish people
}
\examples{
clean_people_names("Ari, Jouni, EsaJokutie 4, pikitie 7, 00500 Helsinki")
x <- seq(0, 10, by = 0.01)
format_num(x)
}
\keyword{clean,}
\keyword{format,}
\keyword{numeric}
\keyword{strees}
\keyword{string,}
