% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hclust.R
\name{plot_hclust}
\alias{plot_hclust}
\title{Custom dendrogram plot}
\usage{
plot_hclust(
  data,
  cut_num = 5,
  clust_var = TRUE,
  horiz = FALSE,
  labels_track_height = 120,
  main = "",
  ...
)
}
\arguments{
\item{data}{Choose the dataframe.}

\item{cut_num}{Choose the number where to cut the number of clusters. Default: 5.}

\item{clust_var}{Choose if you want to cluster variables. Default: TRUE.}

\item{horiz}{Plot horizontal (TRUE) or vertical (FALSE). Default: FALSE.}

\item{labels_track_height}{Make extra space for variable names. Default: 120.}

\item{main}{Title for the plot.}

\item{...}{Pass other parameters to fviz_dend like type, cex, ggtheme.}
}
\description{
Make a custom dendrogram plot.
}
\examples{
plot_hclust(mtcars, cut_num = 4, horiz = TRUE, labels_track_height = 1)
}
\keyword{cluster}
\keyword{data}
\keyword{dendrogram}
