% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing.R
\name{plot_missing}
\alias{plot_missing}
\title{plot_missing}
\usage{
plot_missing(data, file = NULL)
}
\arguments{
\item{data}{Enter your data frame.}

\item{file}{Output file, if you want to save the plot.}
}
\description{
This function helps you to ggplot missing values of a data frame.
}
\examples{
mtcars[10:17, 2:5] <- NA
mtcars[3:5, 8:9] <- NA
plot_missing(mtcars)
}
\keyword{ggplot2}
\keyword{missing}
