% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multibars.R
\name{plot_multibars}
\alias{plot_multibars}
\title{Multiple custom bar plots}
\usage{
plot_multibars(data, row = 1:nrow(data), labels = waiver())
}
\arguments{
\item{data}{dataframe}

\item{row}{select which row to plot. Default: 1:nrow(.)}

\item{labels}{select labels. Default: waiver()}
}
\description{
Create a multiple custom bar plots. With purr::map you can create a list of barplots.
}
\examples{
# Plot multiple bar plots to same grid
library(gridExtra)

plot_basic <- mtcars \%>\%
  dplyr::group_by(cyl) \%>\%
  dplyr::summarise_at(dplyr::vars(mpg, disp, hp, wt), ~mean(.)) \%>\%
  tidyr::gather(2:ncol(.), key = "question", value = "values") \%>\%
  tidyr::spread(cyl, values) \%>\%
  {purrr::map(list(1, 2, 3), function(x) plot_multibars(., x))}
do.call("grid.arrange", c(plot_basic, ncol = 2))

# Plot a single bar plot
plot_multibars(mtcars, row = 3)
}
\keyword{ggplot2}
\keyword{multiplot}
\keyword{plot}
