% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wordcloud.R
\name{plot_wordcloud}
\alias{plot_wordcloud}
\title{plot_wordcloud}
\usage{
plot_wordcloud(x, output = "./wordcloud.png", scale = c(4, 0.5),
  removePunctuation = TRUE, removeNumbers = TRUE,
  stripWhitespace = TRUE, colors = ilmarinen_cols())
}
\arguments{
\item{x}{Vector of sentences or words.}

\item{output}{Output path.}

\item{scale}{A vector of length 2 indicating the range of the size of the words.}

\item{removePunctuation}{Remove punctuation. Default: TRUE.}

\item{removeNumbers}{Remove numbers. Default: TRUE.}

\item{stripWhitespace}{Remove whitespace. Default: TRUE.}

\item{colors}{Colors to use in wordcloud.}
}
\description{
Make a custom wordcloud plot.
}
\examples{
# Create a vector of month names
x <- c()
for (i in 1:12) {
 x <- append(x, rep(month.name[i], i * 2))
}
plot_wordcloud(x, "./wordcloud.png")
}
\keyword{plot,}
\keyword{wordcloud}
