% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wordcloud.R
\name{plot_wordcloud}
\alias{plot_wordcloud}
\title{plot_wordcloud}
\usage{
plot_wordcloud(
  x,
  output = "./wordcloud.png",
  scale = c(4, 0.5),
  width = 12,
  height = 7,
  units = "in",
  res = 150,
  fixed.asp = FALSE,
  removePunctuation = TRUE,
  removeNumbers = TRUE,
  stripWhitespace = TRUE,
  colors = brand_cols()
)
}
\arguments{
\item{x}{Vector of sentences or words.}

\item{output}{Output path.}

\item{scale}{A vector of length 2 indicating the range of the size of the words.}

\item{res}{The nominal resolution in ppi which will be recorded in the bitmap
file, if a positive integer. Also used for units other than the default.
If not specified, taken as 72 ppi to set the size of text and line widths.}

\item{fixed.asp}{if TRUE, the aspect ratio is fixed. Variable aspect ratio
only supported if rot.per==0}

\item{removePunctuation}{Remove punctuation. Default: TRUE.}

\item{removeNumbers}{Remove numbers. Default: TRUE.}

\item{stripWhitespace}{Remove whitespace. Default: TRUE.}

\item{colors}{Colors to use in wordcloud.}
}
\description{
Make a custom wordcloud plot.
}
\examples{
# Create a vector of month names
x <- c()
for (i in 1:12) {
 x <- append(x, rep(month.name[i], i * 2))
}
plot_wordcloud(x, "./wordcloud.png")
}
\keyword{plot,}
\keyword{wordcloud}
