% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questback_integration.R
\name{qb_get_responses}
\alias{qb_get_responses}
\title{qb_get_responses}
\usage{
qb_get_responses(
  filename,
  quest_id,
  sid,
  username,
  password,
  last_updated = "12 hours",
  latest_days = NULL
)
}
\arguments{
\item{filename}{Filename to save the imported data.}

\item{quest_id}{Quest id from Questback Essentials.}

\item{sid}{Quest security id from Questback Essentials.}

\item{username}{Username for Questback Essentials.}

\item{password}{Password for Questback Essentials.}

\item{last_updated}{If time is less than chosen, then do not execute.
Possible time choises are secs, mins, hours, days and weeks. Default is 12 hours.}

\item{latest_days}{Not yet implemented, because IntegrationUtility demands
dates to be in form YYYY-MM-DD.}
}
\description{
Get responses to a survey from Questback Essentials.
}
\keyword{essentials,}
\keyword{questback}
\keyword{survey}
