% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaces.R
\name{replaces}
\alias{replaces}
\title{Replace corresponding elements}
\usage{
replaces(x, from, to, last = FALSE, evaluate = TRUE)
}
\arguments{
\item{x}{vector}

\item{from}{the replaced vector}

\item{to}{the replacing vector}

\item{last}{the last value}

\item{evaluate}{if TRUE, the replaced x will be returned, else, the call sentence before evaluation}
}
\value{
replaced x
}
\description{
Replace corresponding elements
}
\examples{
replaces(1:10, 1:3, 3:5)
}

