% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setProxy.R
\name{setProxy}
\alias{setProxy}
\title{Undescribed}
\usage{
setProxy(url = getOption("jmisc.proxy.url"),
  port = getOption("jmisc.proxy.port"))
}
\arguments{
\item{Undescribed}{}
}
\value{
Undescribed
}
\description{
Undescribed
}

