\name{sourceAll}
\alias{sourceAll}
\title{Source all the R files of a directory}
\usage{
sourceAll(path = ".", ...)
}
\arguments{
  \item{path}{path of the directory}

  \item{...}{other arguments to source}
}
\description{
Source all file with extension .r or .R
}
\examples{
\dontrun{
sourceAll()
}
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}
\seealso{
\link{repCol}
}

